Refactored app package: sample to partagix
