+++ app/src/main/AndroidManifest.xml
@@ -11,18 +11,18 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.SampleApp"
+        android:theme="@style/Theme.Partagix"
         tools:targetApi="31">
         <activity
-            android:name=".SecondActivity"
+            android:name="com.android.partagix.SecondActivity"
             android:exported="false"
             android:label="@string/title_activity_second"
-            android:theme="@style/Theme.SampleApp" />
+            android:theme="@style/Theme.Partagix" />
         <activity
-            android:name=".MainActivity"
+            android:name="com.android.partagix.MainActivity"
             android:exported="true"
             android:label="@string/title_activity_main"
-            android:theme="@style/Theme.SampleApp">
+            android:theme="@style/Theme.Partagix">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
