+++ app/src/main/java/com/android/partagix/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.android.partagix.ui.theme
 
 import android.app.Activity
 import android.os.Build
@@ -34,7 +34,7 @@ private val LightColorScheme =
         )
 
 @Composable
-fun SampleAppTheme(
+fun PartagixAppTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
