+++ app/src/test/java/com/android/partagix/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.android.partagix
 
 import org.junit.Assert.*
 import org.junit.Test
